from_learn_go_with_tests/mocking/countdown_test.go
package main

import (
	"bytes"
	"reflect"
	"testing"
)

type SpySleeper struct {
	Calls int
}

type SpyCountdownOperations struct {
	Calls []string
}

// For mocking test
const write = "write"
const sleep = "sleep"

func (s *SpyCountdownOperations) Sleep() {
	s.Calls = append(s.Calls, sleep)
}

func (s *SpyCountdownOperations) Write(p []byte) (int, error) {
	s.Calls = append(s.Calls, write)
	return 0, nil
}

func TestCountdown(t *testing.T) {
	t.Run("print 3 to Go!", func(t *testing.T) {
		buffer := &bytes.Buffer{}
		spySleeper := &SpyCountdownOperations{}

		// We can send 'buffer' there since it's Buffer struct that has 'Write' Method
		// like 'Write' in io.Writer interface
		Countdown(buffer, spySleeper)

		got := buffer.String()
		want := `3
2
1
Go!`

		if got != want {
			t.Errorf("got %q want %q", got, want)
		}
	})

	t.Run("sleep before every print", func(t *testing.T) {
		spySleepPrinter := &SpyCountdownOperations{}
		Countdown(spySleepPrinter, spySleepPrinter)

		wantOrderOfExecution := []string{
			write,
			sleep,
			write,
			sleep,
			write,
			sleep,
			write,
		}

		if !reflect.DeepEqual(wantOrderOfExecution, spySleepPrinter.Calls) {
			t.Errorf("wanted calls %v got %v", wantOrderOfExecution, spySleepPrinter.Calls)
		}
	})
}
